Move FindInFiles tests into a separate file
