+++ test/UnitTestSuite.js
@@ -48,6 +48,7 @@ define(function (require, exports, module) {
     require("spec/FileFilters-test");
     require("spec/FileSystem-test");
     require("spec/FileUtils-test");
+    require("spec/FindInFiles-test");
     require("spec/FindReplace-test");
     require("spec/HTMLInstrumentation-test");
     require("spec/HTMLSimpleDOM-test");
